Split motor files; moved to match names
