+++ Grbl_Esp32/src/Motors/Motor.cpp
@@ -31,8 +31,14 @@
         TMC2130 Datasheet https://www.trinamic.com/fileadmin/assets/Products/ICs_Documents/TMC2130_datasheet.pdf
 */
 
+#include "Motor.h"
 #include "../grbl.h"
-#include "MotorClass.h"
+
+#include "NullMotor.h"
+#include "StandardStepper.h"
+#include "UnipolarMotor.h"
+#include "RcServo.h"
+#include "TrinamicDriver.h"
 
 Motor*              myMotor[MAX_AXES][MAX_GANGED];  // number of axes (normal and ganged)
 static TaskHandle_t readSgTaskHandle      = 0;      // for realtime stallguard data diaplay
@@ -442,29 +448,29 @@ void TMC2130Stepper::switchCSpin(bool state) {
 
 // ============================== Class Methods ================================================
 
-Motor ::Motor() {
+Motor::Motor() {
     type_id = MOTOR;
 }
 
-void Motor ::init() {
+void Motor::init() {
     _homing_mask = 0;
 }
 
-void Motor ::config_message() {}
-void Motor ::debug_message() {}
-void Motor ::read_settings() {}
-void Motor ::set_disable(bool disable) {}
-void Motor ::set_direction_pins(uint8_t onMask) {}
-void Motor ::step(uint8_t step_mask, uint8_t dir_mask) {}
-bool Motor ::test() {
+void Motor::config_message() {}
+void Motor::debug_message() {}
+void Motor::read_settings() {}
+void Motor::set_disable(bool disable) {}
+void Motor::set_direction_pins(uint8_t onMask) {}
+void Motor::step(uint8_t step_mask, uint8_t dir_mask) {}
+bool Motor::test() {
     return true;
 };  // true = OK
-void Motor ::update() {}
+void Motor::update() {}
 
-void Motor ::set_axis_name() {
+void Motor::set_axis_name() {
     sprintf(_axis_name, "%c%s", report_get_axis_letter(axis_index), dual_axis_index ? "2" : "");
 }
 
-void Motor ::set_homing_mode(uint8_t homing_mask, bool isHoming) {
+void Motor::set_homing_mode(uint8_t homing_mask, bool isHoming) {
     _homing_mask = homing_mask;
 }
