+++ Grbl_Esp32/src/Motors/NullMotor.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "Motor.h"
+
+class Nullmotor : public Motor {};
